feat: move sitemap stuff to a component, use its own DAO (closes #1456)
