+++ backend/src/main/java/io/papermc/hangar/components/sitemap/model/SitemapProject.java
@@ -1,0 +1,4 @@
+package io.papermc.hangar.components.sitemap.model;
+
+public record SitemapProject(String slug, long id) {
+}
