+++ backend/src/main/java/io/papermc/hangar/components/sitemap/model/SitemapUser.java
@@ -1,0 +1,4 @@
+package io.papermc.hangar.components.sitemap.model;
+
+public record SitemapUser(String name, long id) {
+}
