+++ backend/src/main/java/io/papermc/hangar/components/sitemap/model/SitemapVersion.java
@@ -1,0 +1,5 @@
+package io.papermc.hangar.components.sitemap.model;
+
+public record SitemapVersion(long id, String versionString, String channel) {
+
+}
